net: inet_is_local_reserved_port() port arg should be unsigned short (bsc#1154353).
suse-commit: 80f18fe2b61d3d1d6239938096e1e107775fde83

+++ include/net/ip.h
@@ -338,7 +338,7 @@ static inline u64 snmp_fold_field64(void __percpu *mib, int offt, size_t syncp_o
 void inet_get_local_port_range(struct net *net, int *low, int *high);
 
 #ifdef CONFIG_SYSCTL
-static inline bool inet_is_local_reserved_port(struct net *net, int port)
+static inline bool inet_is_local_reserved_port(struct net *net, unsigned short port)
 {
 	if (!net->ipv4.sysctl_local_reserved_ports)
 		return false;
@@ -356,7 +356,7 @@ static inline int inet_prot_sock(struct net *net)
 }
 
 #else
-static inline bool inet_is_local_reserved_port(struct net *net, int port)
+static inline bool inet_is_local_reserved_port(struct net *net, unsigned short port)
 {
 	return false;
 }
